feat(filter): implement dynamic query filtering strategy with LLM classification
- Added a new configuration option for filter strategy in `application.conf`.
- Introduced `QueryFilterStrategy` interface and its implementations: `LLMClassificationFilterStrategy` and `NoFilterStrategy`.
- Updated `QueryFilterService` to utilize the selected filter strategy based on configuration.
- Created a factory for generating filter strategies dynamically.
