+++ backend/src/main/resources/application.conf
@@ -57,6 +57,7 @@ gitinsp {
 
   rag {
     use-dynamic-filter = true # checks whether to retrieve code using specific languages (i.e. python, java, c++, scala)
+    filter-strategy = "llm" # Available options: "llm", "none"
     use-conditional-rag = true # asks a model whether to retrieve code from the database or skip it
     model = "gemma3" # the model used to query conditional RAG and dynamic filter
   }
